+++ gazebo/msgs/msgs_TEST.cc
@@ -206,6 +206,67 @@ TEST_F(MsgsTest, ConvertCommonTimeToMsgs)
   EXPECT_EQ(123, v.nsec);
 }
 
+TEST_F(MsgsTest, ConvertMathInertialToMsgs)
+{
+  const auto pose = ignition::math::Pose3d(
+    ignition::math::Vector3d(5, 6, 7),
+    ignition::math::Vector3d(0.4, 0.5, 0.6));
+  msgs::Inertial msg = msgs::Convert(
+      ignition::math::Inertiald(
+        ignition::math::MassMatrix3d(12.0,
+          ignition::math::Vector3d(2, 3, 4),
+          ignition::math::Vector3d(0.1, 0.2, 0.3)),
+        pose));
+
+  EXPECT_DOUBLE_EQ(12.0, msg.mass());
+  EXPECT_DOUBLE_EQ(2.0, msg.ixx());
+  EXPECT_DOUBLE_EQ(3.0, msg.iyy());
+  EXPECT_DOUBLE_EQ(4.0, msg.izz());
+  EXPECT_DOUBLE_EQ(0.1, msg.ixy());
+  EXPECT_DOUBLE_EQ(0.2, msg.ixz());
+  EXPECT_DOUBLE_EQ(0.3, msg.iyz());
+  EXPECT_EQ(pose, msgs::ConvertIgn(msg.pose()));
+}
+
+TEST_F(MsgsTest, ConvertMsgsInertialToMath)
+{
+  const auto pose = ignition::math::Pose3d(
+    ignition::math::Vector3d(5, 6, 7),
+    ignition::math::Vector3d(0.4, 0.5, 0.6));
+  msgs::Inertial msg = msgs::Convert(
+      ignition::math::Inertiald(
+        ignition::math::MassMatrix3d(12.0,
+          ignition::math::Vector3d(2, 3, 4),
+          ignition::math::Vector3d(0.1, 0.2, 0.3)),
+        pose));
+  auto inertial = msgs::Convert(msg);
+
+  EXPECT_DOUBLE_EQ(12.0, inertial.MassMatrix().Mass());
+  EXPECT_DOUBLE_EQ(2.0, inertial.MassMatrix().IXX());
+  EXPECT_DOUBLE_EQ(3.0, inertial.MassMatrix().IYY());
+  EXPECT_DOUBLE_EQ(4.0, inertial.MassMatrix().IZZ());
+  EXPECT_DOUBLE_EQ(0.1, inertial.MassMatrix().IXY());
+  EXPECT_DOUBLE_EQ(0.2, inertial.MassMatrix().IXZ());
+  EXPECT_DOUBLE_EQ(0.3, inertial.MassMatrix().IYZ());
+  EXPECT_EQ(pose, inertial.Pose());
+}
+
+TEST_F(MsgsTest, ConvertMathMassMatrix3ToMsgs)
+{
+  msgs::Inertial msg = msgs::Convert(
+      ignition::math::MassMatrix3d(12.0,
+        ignition::math::Vector3d(2, 3, 4),
+        ignition::math::Vector3d(0.1, 0.2, 0.3)));
+
+  EXPECT_DOUBLE_EQ(12.0, msg.mass());
+  EXPECT_DOUBLE_EQ(2.0, msg.ixx());
+  EXPECT_DOUBLE_EQ(3.0, msg.iyy());
+  EXPECT_DOUBLE_EQ(4.0, msg.izz());
+  EXPECT_DOUBLE_EQ(0.1, msg.ixy());
+  EXPECT_DOUBLE_EQ(0.2, msg.ixz());
+  EXPECT_DOUBLE_EQ(0.3, msg.iyz());
+}
+
 TEST_F(MsgsTest, ConvertMathPlaneToMsgs)
 {
   msgs::PlaneGeom msg = msgs::Convert(
@@ -367,6 +428,46 @@ TEST_F(MsgsTest, SetTime)
   msgs::Set(&msg, common::Time(2, 123));
   EXPECT_EQ(2, msg.sec());
   EXPECT_EQ(123, msg.nsec());
+}
+
+TEST_F(MsgsTest, SetInertial)
+{
+  const auto pose = ignition::math::Pose3d(
+    ignition::math::Vector3d(5, 6, 7),
+    ignition::math::Vector3d(0.4, 0.5, 0.6));
+  msgs::Inertial msg;
+  msgs::Set(&msg, ignition::math::Inertiald(
+      ignition::math::MassMatrix3d(
+        12.0,
+        ignition::math::Vector3d(2, 3, 4),
+        ignition::math::Vector3d(0.1, 0.2, 0.3)),
+      pose));
+
+  EXPECT_DOUBLE_EQ(12.0, msg.mass());
+  EXPECT_DOUBLE_EQ(2.0, msg.ixx());
+  EXPECT_DOUBLE_EQ(3.0, msg.iyy());
+  EXPECT_DOUBLE_EQ(4.0, msg.izz());
+  EXPECT_DOUBLE_EQ(0.1, msg.ixy());
+  EXPECT_DOUBLE_EQ(0.2, msg.ixz());
+  EXPECT_DOUBLE_EQ(0.3, msg.iyz());
+  EXPECT_EQ(pose, msgs::ConvertIgn(msg.pose()));
+}
+
+TEST_F(MsgsTest, SetMassMatrix3)
+{
+  msgs::Inertial msg;
+  msgs::Set(&msg, ignition::math::MassMatrix3d(
+        12.0,
+        ignition::math::Vector3d(2, 3, 4),
+        ignition::math::Vector3d(0.1, 0.2, 0.3)));
+
+  EXPECT_DOUBLE_EQ(12.0, msg.mass());
+  EXPECT_DOUBLE_EQ(2.0, msg.ixx());
+  EXPECT_DOUBLE_EQ(3.0, msg.iyy());
+  EXPECT_DOUBLE_EQ(4.0, msg.izz());
+  EXPECT_DOUBLE_EQ(0.1, msg.ixy());
+  EXPECT_DOUBLE_EQ(0.2, msg.ixz());
+  EXPECT_DOUBLE_EQ(0.3, msg.iyz());
 }
 
 TEST_F(MsgsTest, SetPlane)